Fix #34: Migrasi monetary columns ke integer cents (#69)
- Migration: decimal(12,2) → bigInteger, konversi data ×100
- Models: casts decimal:2 → integer
- Form Requests: prepareForValidation() konversi input rupiah → cents
- Controllers: hapus (float) casts dan bccomp()
- Views: display / 100
- Tests/Factories: form submission rupiah, stored assertions cents
- Fix migration #23: drop index sebelum drop column room_name (SQLite compat)
- Fix pre-existing test failures akibat room_name column di-drop
- Add room_name validation to StoreSantriRequest & UpdateSantriRequest

+++ app/Actions/Santri/GenerateMonthlySantriInvoices.php
@@ -21,7 +21,7 @@ class GenerateMonthlySantriInvoices
         int $periodMonth,
         int $periodYear,
         string $dueDate,
-        float $amount,
+        int $amount,
         ?string $notes = null,
         ?int $createdBy = null,
         bool $dryRun = false
